Define use case request & response as dto
