+++ backend/app/schemas/interview_schema.py
@@ -1,47 +1,61 @@
-from app.domain.entities.difficulty import Difficulty
+# generated by fastapi-codegen:
+#   filename:  api.yml
+#   timestamp: 2025-05-19T12:37:04+00:00
+
+from __future__ import annotations
+
+from enum import Enum
+from typing import List
+
 from pydantic import BaseModel, Field
 
 
-class SetUpInterviewRequest(BaseModel):
+# TODO: リファクタリング後に削除する
+class Difficulty(Enum):
+    easy = "easy"
+    normal = "normal"
+    hard = "hard"
+    extreme = "extreme"
+
+
+# TODO: リファクタリング後に削除する
+class InterviewPostRequest(BaseModel):
     source_code: bytes
     difficulty: Difficulty
     total_question: int = Field(gt=0)
 
 
-class SetUpInterviewResponse(BaseModel):
+# TODO: リファクタリング後に削除する
+class InterviewPostResponse(BaseModel):
     interview_id: str
     first_question: str
 
 
-class GetFeedbackRequest(BaseModel):
-    interview_id: str
-    question_id: str
+class InterviewInterviewIdPostRequest(BaseModel):
+    question_id: int
     message: str
 
 
-class GetFeedbackResponse(BaseModel):
-    interview_id: str
-    question_id: str
-    score: int
-    comment: str
+class InterviewInterviewIdPostResponse(BaseModel):
+    question_id: int
+    score: int = Field(ge=0, le=100)
+    response: str
+    continue_question: bool
 
 
-class GetQuestionRequest(BaseModel):
-    interview_id: str
-    question_id: str
+class InterviewInterviewIdPostErrorResponse(BaseModel):
+    error_message: str
 
 
-class GetQuestionResponse(BaseModel):
-    interview_id: str
-    question_id: str
+class InterviewInterviewIdGetResponse(BaseModel):
+    question_id: int
     question: str
 
 
-class GetInterviewResultRequest(BaseModel):
-    interview_id: str
+class InterviewInterviewIdResultGetResponse(BaseModel):
+    scores: List[int]
+    general_review: str
 
 
-class GetInterviewResultResponse(BaseModel):
-    interview_id: str
-    scores: list[int]
-    overall_review: str
+class InterviewInterviewIdResultGetErrorResponse(BaseModel):
+    error_message: str
