+++ backend/app/usecase/usecases/get_feedback_usecase.py
@@ -4,7 +4,7 @@ from app.domain.entities.chat_history import ChatMessage
 from app.domain.llm_clients.llm_client import LLMClient
 from app.domain.repositories.interview_repository import InterviewRepository
 from app.domain.repositories.source_code_repository import SourceCodeRepository
-from app.schemas.interview_schema import GetFeedbackRequest, GetFeedbackResponse
+from app.usecase.dtos.interview_dto import GetFeedbackRequest, GetFeedbackResponse
 
 
 class GetFeedbackUseCase:
