+++ backend/app/usecase/usecases/get_interview_result_usecase.py
@@ -1,6 +1,6 @@
 from app.domain.llm_clients.llm_client import LLMClient
 from app.domain.repositories.interview_repository import InterviewRepository
-from app.schemas.interview_schema import (
+from app.usecase.dtos.interview_dto import (
     GetInterviewResultRequest,
     GetInterviewResultResponse,
 )
